Added scrollview in keyboard in Personal data
